+++ macros/monsterFeatures/sleep.js
@@ -17,7 +17,7 @@ for (const target of lastArg.targets) {
     duration: { rounds: 10, seconds: 60, startRound: gameRound, startTime: game.time.worldTime },
     flags: { dae: { specialDuration: ["isDamaged"] } },
     changes: [
-      { key: "macro.CE", mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM, value: "Unconscious", priority: 20 },
+      DDBImporter.EffectHelper.generateCEStatusEffectChange("Unconscious", 20, true),
     ]
   };