+++ macros/spells/sleep.js
@@ -36,8 +36,7 @@ for (let target of targets) {
       duration: { rounds: 10, seconds: 60, startRound: gameRound, startTime: game.time.worldTime },
       flags: { dae: { specialDuration: ["isDamaged"] } },
       changes: [
-        // { key: "macro.CE", mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM, value: "Prone", priority: 20 },
-        { key: "macro.CE", mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM, value: "Unconscious", priority: 20 },
+        DDBImporter.EffectHelper.generateCEStatusEffectChange(game.i18n.localize("Unconscious"), 20, true),
       ]
     };